+++ modules/mgr/chatmgr.js
@@ -158,6 +158,7 @@ function getSpecialMoveFlavor(resp) {
     if (specialMove === SPECIAL.JAB) mods.push(game.i18n.localize('HM.jab'));
     if (specialMove === SPECIAL.SCAMPER) mods.push(game.i18n.localize('EFFECT.scamper'));
     if (specialMove === SPECIAL.SET4CHARGE) mods.push(game.i18n.localize('HM.specSelect.s4c'));
+    if (specialMove === SPECIAL.WITHDRAWL) mods.push(game.i18n.localize('HM.specSelect.wdrawl'));
     if (shieldHit) mods.push(game.i18n.localize('HM.blocked'));
     return mods.length ? ` (${mods.join(', ')})` : '';
 }