+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [Unreleased]
+### Added
+- Fighting Withdrawl special move.
+
 ## [0.2.19] - 2022-10-09
 ### Added
 - Compendium items.